Added specific covariance parameter to KMeansInputGenerator

+++ peel-flink-kmeans-datagens/src/main/scala/de/tu_berlin/dima/bdapro/datagen/flink/Distributions.scala
@@ -5,7 +5,7 @@ import org.apache.commons.math3.random.Well19937c
 
 object Distributions {
 
-  case class ContinousUniform(upper: Int, lower: Int = 0) {
+  case class ContinousUniform(upper: Double, lower: Double = 0) {
     val distribution = new UniformRealDistribution(lower, upper)
 
     def sample(cp: Double) = distribution.inverseCumulativeProbability(cp)
